@@ -4,6 +4,6 @@ export default {
   out: './drizzle',
   driver: 'pg',
   dbCredentials: {
-    connectionString: `postgres://${process.env.DB_USERNAME}:${process.env.DB_USERNAME}@localhost:5432/${process.env.DB_NAME}`,
+    connectionString: `postgres://${process.env.DB_USERNAME}:${process.env.DB_USERNAME}@postgres:5432/${process.env.DB_NAME}`,
   },
 } satisfies Config
