@@ -1,4 +1,4 @@
-import { FastifyPluginAsync } from "fastify";
+import { FastifyPluginAsync } from 'fastify'
 import {
   createRole,
   getRolesPaginate,
@@ -6,7 +6,7 @@ import {
   updateRoleById,
   deleteRole,
   getRoleWithPermissions,
-} from "../../services/roleService";
+} from '../../services/roleService.js'
 import {
   CreateRoleSchema,
   CreateRoleSchemaType,
@@ -16,11 +16,11 @@ import {
   PatchRoleSchemaType,
   getRoleByIdSchema,
   getRoleByIdType,
-} from "./roleSchema";
+} from './roleSchema'
 
 const roles: FastifyPluginAsync = async (fastify, opts): Promise<void> => {
   fastify.get<{ Querystring: ListRoleQueryParamSchemaType }>(
-    "/",
+    '/',
     {
       onRequest: [fastify.authenticate],
       schema: {
@@ -28,25 +28,17 @@ const roles: FastifyPluginAsync = async (fastify, opts): Promise<void> => {
       },
     },
     async function (request, reply) {
-      let { search = "", limit = 10, page = 1 } = request.query;
-      const offset = fastify.findOffset(limit, page);
-      const result = await getRolesPaginate(search, limit, page, offset);
-      let message: string = fastify.responseMessage(
-        "roles",
-        result.data.length
-      );
-      let requestUrl: string | null =
-        request.protocol + "://" + request.hostname + request.url;
-      const nextUrl: string | null = fastify.findNextPageUrl(
-        requestUrl,
-        result.totalPage,
-        page
-      );
+      let { search = '', limit = 10, page = 1 } = request.query
+      const offset = fastify.findOffset(limit, page)
+      const result = await getRolesPaginate(search, limit, page, offset)
+      let message: string = fastify.responseMessage('roles', result.data.length)
+      let requestUrl: string | null = request.protocol + '://' + request.hostname + request.url
+      const nextUrl: string | null = fastify.findNextPageUrl(requestUrl, result.totalPage, page)
       const previousUrl: string | null = fastify.findPreviousPageUrl(
         requestUrl,
         result.totalPage,
-        page
-      );
+        page,
+      )
 
       return {
         message: message,
@@ -57,11 +49,11 @@ const roles: FastifyPluginAsync = async (fastify, opts): Promise<void> => {
         page: page,
         limit: limit,
         data: result.data,
-      };
-    }
-  );
+      }
+    },
+  )
   fastify.post<{ Body: CreateRoleSchemaType; Reply: object }>(
-    "/",
+    '/',
     {
       onRequest: [fastify.authenticate],
       schema: {
@@ -70,13 +62,13 @@ const roles: FastifyPluginAsync = async (fastify, opts): Promise<void> => {
       },
     },
     async (request, reply) => {
-      const { roleName, description = "" } = request.body;
-      const role = await createRole(roleName, description);
-      reply.status(201).send({ message: "Role created", data: role });
-    }
-  );
+      const { roleName, description = '' } = request.body
+      const role = await createRole(roleName, description)
+      reply.status(201).send({ message: 'Role created', data: role })
+    },
+  )
   fastify.get<{ Params: getRoleByIdType }>(
-    "/:id",
+    '/:id',
     {
       onRequest: [fastify.authenticate],
       schema: {
@@ -84,21 +76,21 @@ const roles: FastifyPluginAsync = async (fastify, opts): Promise<void> => {
       },
     },
     async (request, reply) => {
-      const id = request.params.id;
-      const role = await getRoleById(id);
+      const id = request.params.id
+      const role = await getRoleById(id)
       if (role == null) {
-        return reply.status(404).send({ message: "role not found" });
+        return reply.status(404).send({ message: 'role not found' })
       }
-      const rolePermissions = await getRoleWithPermissions(role.id);
-      reply.status(200).send({ role: role, permissions: rolePermissions });
-    }
-  );
+      const rolePermissions = await getRoleWithPermissions(role.id)
+      reply.status(200).send({ role: role, permissions: rolePermissions })
+    },
+  )
   fastify.patch<{
-    Body: PatchRoleSchemaType;
-    Reply: object;
-    Params: getRoleByIdType;
+    Body: PatchRoleSchemaType
+    Reply: object
+    Params: getRoleByIdType
   }>(
-    "/:id",
+    '/:id',
     {
       onRequest: [fastify.authenticate],
       schema: {
@@ -107,23 +99,21 @@ const roles: FastifyPluginAsync = async (fastify, opts): Promise<void> => {
       },
     },
     async (request, reply) => {
-      const roleData = request.body;
+      const roleData = request.body
       if (Object.keys(roleData).length == 0) {
-        return reply
-          .status(422)
-          .send({ message: "Provide at least one column to update." });
+        return reply.status(422).send({ message: 'Provide at least one column to update.' })
       }
-      const id = request.params.id;
+      const id = request.params.id
 
-      const role = await updateRoleById(id, roleData);
+      const role = await updateRoleById(id, roleData)
       if (role.length == 0) {
-        return reply.status(404).send({ message: "role not found" });
+        return reply.status(404).send({ message: 'role not found' })
       }
-      reply.status(201).send({ message: "Role Updated", data: role });
-    }
-  );
+      reply.status(201).send({ message: 'Role Updated', data: role })
+    },
+  )
   fastify.delete<{ Params: getRoleByIdType }>(
-    "/:id",
+    '/:id',
     {
       onRequest: [fastify.authenticate],
       schema: {
@@ -131,14 +121,14 @@ const roles: FastifyPluginAsync = async (fastify, opts): Promise<void> => {
       },
     },
     async (request, reply) => {
-      const id = request.params.id;
-      const deletedRole = await deleteRole(id);
+      const id = request.params.id
+      const deletedRole = await deleteRole(id)
       if (deletedRole == null) {
-        return reply.status(404).send({ message: "Role doesn't exist!" });
+        return reply.status(404).send({ message: "Role doesn't exist!" })
       }
-      return reply.status(200).send({ message: "Role deleted" });
-    }
-  );
-};
+      return reply.status(200).send({ message: 'Role deleted' })
+    },
+  )
+}
 
-export default roles;
+export default roles
